refactor: 리듀서 네이밍 변경 - customer -> user로 변경
