+++ src/redux/store.ts
@@ -2,7 +2,7 @@ import { createStore, combineReducers, applyMiddleware } from 'redux';
 import ReduxThunk from 'redux-thunk';
 import logger from 'redux-logger';
 import cartReducer from 'redux/modules/cart';
-import customerReducer from 'redux/modules/customer';
+import userReducer from 'redux/modules/user';
 import productsReducer from 'redux/modules/products';
 import snackBarReducer from 'redux/modules/snackBar';
 
@@ -11,7 +11,7 @@ export type RootState = ReturnType<typeof store.getState>;
 
 const rootReducer = combineReducers({
   cart: cartReducer,
-  customer: customerReducer,
+  user: userReducer,
   products: productsReducer,
   snackBar: snackBarReducer,
 });
